♻️ Encode the challenge on-chain
You may probably ask why we encode the challenge in base64 on-chain
instead of of sending it already encoded to save some gas. This library
is opinionated and assumes that it is used in the context of Account
Abstraction. In this context, valuable information required to proceed
with the transaction correctly will be stored in the challenge field
meaning we need the challenge in clear to use later in the flow. That's
why we decided to encode it here.