💡 explain why the overflow check is not needed
During the challenge hashing step, we add the challenge offset to the
length of the challenge encoded to get the end index of the challenge
in the client data. This is a potential overflow point, but it is cover
one step later because we use the end index to slice the client data. If
the end index overflow, it will be in that case lower than the start index.
In that case, the EVM automatically raise an Error.
Keep in mind the start index is equals to the challenge offset, this is
an important point to understand why the overflow check is not needed. In
the case the start index is calculated using data not correlated to the
calculus of the end index, the overflow check is needed.